Refactor Nation and Personality to single-class per file, UI separation, prepare Pedia visibility of Personality
